backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
supabase==2.11.0
google-genai
httpx==0.28.1
python-multipart==0.0.20
pydantic==2.10.4
PyPDF2
pandas
openpyxl
PyJWT>=2.8.0
redis>=5.0.0
pytest>=8.0.0
